deuxfleurs

Type Alias InitialState

pub type InitialState = State<InnerBareState>;

Starting point to build the app.

Aliased Type

pub struct InitialState(/* private fields */);

Implementations

impl InitialState

pub fn run<S: Into<String>>(self, width: u32, height: u32, id: Option<S>)

Show the window and start the app.

In wasm, width and height are ignored and css is used to define the dimensions (allowing for dimensions in % and vh/vw).

Arguments:

width: requested width of the app (no effect in wasm)

height: requested height of the app (no effect in wasm)

id: serves as window title, or id shape to attach to. If None uses "State".

use deuxfleurs::load_mesh;

let (spot_v, spot_f) = load_mesh("examples/assets/spot.obj").await.unwrap();
let mut handle = deuxfleurs::init();
handle.register_surface("Spot", spot_v, spot_f);
let mut handle = handle.run(1920, 1080, Some("deuxfleurs"));

pub fn run_headless(self) -> RunningState

Available on non-WebAssembly only.

Run the app without a window. Allows running the app in environment where no display is available and taking screenshots automatically.

Currently only available on non wasm targets, as webGL requires a context.

use deuxfleurs::load_mesh;

let (spot_v, spot_f) = load_mesh("examples/assets/spot.obj").await.unwrap();
let mut handle = deuxfleurs::init();
handle.register_surface("Spot", spot_v, spot_f);
let mut handle = handle.run_headless();
handle.screenshot();

pub fn run_with_callback<S: Into<String>, U: FnMut(&mut Ui, &mut RunningState)>( self, width: u32, height: u32, id: Option<S>, callback: U, )

Specify a callback that will be called once every frame.

Passes an [egui::Ui] and a RunningState arguments which can be used to add UI elements and modify state accordingly.