deuxfleurs::internal

Struct State

pub struct State<T>(/* private fields */);

Implementations

impl<T: StateTrait> State<T>

pub fn register_surface<V: Vertices, I: Into<SurfaceIndices>>( &mut self, name: impl Into<String>, vertices: V, indices: I, ) -> SurfaceMut<'_, T>

Register a new surface. If an existing one with same number of vertices and same faces exists, previous settings and data are recovered.

See Surface and SurfaceMut for how to add data to the created shape.

pub fn get_surface_mut(&mut self, name: &str) -> Option<SurfaceMut<'_, T>>

pub fn get_surface(&self, name: &str) -> Option<&Surface<T>>

pub fn remove_surface(&mut self, name: &str)

pub fn register_point_cloud<V: Vertices>( &mut self, name: impl Into<String>, positions: V, ) -> PointCloudMut<'_, T>

Register a new point cloud. If an existing one with same number of points exists, previous settings and data are recovered.

See PointCloud and PointCloudMut for how to add data to the created shape.

pub fn get_point_cloud_mut( &mut self, name: &str, ) -> Option<PointCloudMut<'_, T>>

pub fn get_point_cloud(&self, name: &str) -> Option<&PointCloud<T>>

pub fn remove_point_cloud(&mut self, name: &str)

pub fn register_segment<V: Vertices>( &mut self, name: impl Into<String>, positions: V, connections: Vec<[u32; 2]>, ) -> SegmentMut<'_, T>

Register a list of segments. If an existing one with same number of points and same connextions exists, previous settings and data are recovered.

See Segment and SegmentMut for how to add data to the created shape.

Arguments :

positions: segments extremities

connections: segments denoted by extremities indices

pub fn get_segment_mut(&mut self, name: &str) -> Option<SegmentMut<'_, T>>

pub fn get_segment(&self, name: &str) -> Option<&Segment<T>>

pub fn remove_segment(&mut self, name: &str)

pub fn load_from_state_file(&mut self, path: impl AsRef<Path>) -> Result<(), ()>

Available on crate feature saves and non-WebAssembly only.

Load app state from given file content.

pub fn load_from_state_slice(&mut self, data: &[u8]) -> Result<(), ()>

Available on crate feature saves only.

Load app state from given buffer.

pub fn save_state_file(&self, path: impl AsRef<Path>) -> Result<(), ()>

Available on crate feature saves and non-WebAssembly only.

Save current state in cbor into chosen file.

pub fn save_state_vec(&self) -> Result<Vec<u8>, ()>

Available on crate feature saves only.

Save current state in cbor into buffer.

pub fn get_settings_mut(&mut self) -> &mut Settings

pub fn get_settings(&self) -> &Settings

pub fn set_camera(&mut self, eye: [f32; 3], target: [f32; 3], up: [f32; 3])

pub fn get_camera(&self) -> ([f32; 3], [f32; 3], [f32; 3])

Result is (eye, target, up).

impl State<InnerBareState>

pub fn run<S: Into<String>>(self, width: u32, height: u32, id: Option<S>)

Show the window and start the app.

In wasm, width and height are ignored and css is used to define the dimensions (allowing for dimensions in % and vh/vw).

Arguments:

width: requested width of the app (no effect in wasm)

height: requested height of the app (no effect in wasm)

id: serves as window title, or id shape to attach to. If None uses "State".

use deuxfleurs::load_mesh;

let (spot_v, spot_f) = load_mesh("examples/assets/spot.obj").await.unwrap();
let mut handle = deuxfleurs::init();
handle.register_surface("Spot", spot_v, spot_f);
let mut handle = handle.run(1920, 1080, Some("deuxfleurs"));

pub fn run_headless(self) -> RunningState

Available on non-WebAssembly only.

Run the app without a window. Allows running the app in environment where no display is available and taking screenshots automatically.

Currently only available on non wasm targets, as webGL requires a context.

use deuxfleurs::load_mesh;

let (spot_v, spot_f) = load_mesh("examples/assets/spot.obj").await.unwrap();
let mut handle = deuxfleurs::init();
handle.register_surface("Spot", spot_v, spot_f);
let mut handle = handle.run_headless();
handle.screenshot();

pub fn run_with_callback<S: Into<String>, U: FnMut(&mut Ui, &mut RunningState)>( self, width: u32, height: u32, id: Option<S>, callback: U, )

Specify a callback that will be called once every frame.

Passes an [egui::Ui] and a RunningState arguments which can be used to add UI elements and modify state accordingly.

impl State<InnerGraphicalState>

pub fn resize_scene(&mut self)

Fit camera and ground to match the visible shapes

pub fn screenshot(&mut self)

Take a screenshot of the scene. The screenshot is then saved on disk under the name screenshot_[nnn].png, with nnn incrementing each time.

pub fn screenshot_to_buffer(&mut self) -> Result<Vec<u8>, ()>

Take a screenshot of the scene. The screenshot is then returned as a vector oy bytes, each storing r g b a (in order) values of each pixel.

Should not fail, unless internal buffer storage is messed up.

pub fn get_picked(&self) -> &Option<(String, Picked)>

Get current selected object: first the name, then index i and type of the selected element

pub fn refresh(&mut self)

Politely ask to render the next frame, even if no change is detected